refactor: move utilities and schedule menu
